+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v3.18.0] - 2026-01-05
+### Added
+- [#274] Support for data volume storage class in Dogu-CR
+
 ## [v3.17.0] - 2025-12-11
 ### Fixed
 - [#273] Recreate component service accounts if dogus are restored from a backup.